btrfs/220: fix clear_cache and inode_cache option tests
I recently changed clear_cache to not appear in mount options, as it has
one shot semantics, which was breaking this test. Test explicitly that
it _doesn't_ appear, which properly fails on old filesystems and passes
on misc-next.

The patch that changed this behavior was:
8b228324a8ce btrfs: clear free space tree on ro->rw remount

Separately, inode_cache is deprecated and will never appear in mount
options; remove it entirely.

Signed-off-by: Boris Burkov