src/idmapped-mounts: use renameat instead of renameat2
We don't add fallback for renameat2, so it build failed on old
kernel.  Also, renameat2() man-pages said as below:

"renameat2() has an additional flags argument.  A renameat2() call
with a zero flags argument is equivalent to renameat()."

In this src file, we use zero flag for all renameat2 call. So we use
renameat directly.

Signed-off-by: Yang Xu <xuyang2018.jy@fujitsu.com>
Reviewed-by: Christian Brauner <christian.brauner@ubuntu.com>
Reviewed-by:Anju T Sudhakar<anju@linux.vnet.ibm.com>
Signed-off-by: Eryu Guan <guaneryu@gmail.com>